% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besag.R
\name{besag_standard}
\alias{besag_standard}
\title{Standardized Besag or ICAR effects}
\usage{
besag_standard(x, adj_mat)
}
\arguments{
\item{x}{An ordered factor, with more than 1 level.}

\item{adj_mat}{Matrix needed for `model="besag"`, representing the adjacency matrix from the lattice.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item{`precision`: precision matrix.}
\item{`basis`:  basis matrix evaluated at `x`}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
 \item{`null_space`: a matrix, representing the null space of the precision matrix.}
 \item{`X_distribution`: a numeric vector sampled from the Uniform distribution on X.}
 \item{`basis_distribution`: basis matrix evaluated at `X_distribution`.}
}
}
\description{
#'@description
`besag_standard()` provides a list of elements to build a standardized Besag effect.
}
