% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline2D.R
\name{bspline_2D}
\alias{bspline_2D}
\title{Create a basis of cubic uniform 2D B-Spline functions}
\usage{
bspline_2D(x, K, m = c(0, 0), M = c(1, 1))
}
\arguments{
\item{x}{A matrix of 2 columns. The first column must contain values between \code{m[1]} and \code{M[1]}. The second column must contain values between \code{m[2]} and \code{M[2]}.}

\item{K}{Two positive integers larger than 3, specifying the number of basis functions on each dimension.}

\item{m, M}{Two pairs of numbers indicating respectively the lower and upper bounds of the rectangular support on which to define the basis functions. By default, \code{c(0,0)} and \code{c(1,1)}.}
}
\value{
A matrix with \code{K1}x\code{K2} columns and a number of rows equal to the length of \code{x1} and \code{x2}.
}
\description{
\code{bspline_2D()} creates a basis matrix with \code{K1}x\code{K2} columns, each of them evaluating the numeric vector \code{x} at a 2-dimensional cubic B-Spline function, defined through the Kronecker product of 2 unidimensional B-Spline functions on the rectangular support delimited by \verb{m,M}.
}
\examples{
x_grid <- as.matrix(expand.grid(
  seq(0, 1, length.out = 100),
  seq(0, 1, length.out = 100)
))

bspline_2D(x_grid, K=c(5,5))

}
