% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_GMRF}
\alias{check_GMRF}
\title{Checks if the effect has been properly standardized.}
\usage{
check_GMRF(realizations, fixed = T, X_dist = NULL)
}
\arguments{
\item{realizations}{A numeric matrix, containing a realization of the process in each column.}

\item{fixed}{Logical, indicating whether the effect must be treated as fixed or random. By default, \code{TRUE}.}

\item{X_dist}{A numeric matrix, containing a realization of the process in each column.}
}
\value{
Plot.
}
\description{
\code{check_GMRF()} returns a plot with 2 panels: the means and variances of the realizations are plotted via histograms. The means of both statistics are also reported, which should be equal respectively to 0 (only if \code{fixed=TRUE}) and 1.
}
