% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBoundaries.R
\name{checkBoundaries}
\alias{checkBoundaries}
\title{Returns the boundaries for the covariates}
\usage{
checkBoundaries(x, m = NULL, M = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{m}{A number indicating the lower boundary of the support of X. If not provided, it is set to the minimum value from `x`.}

\item{M}{A number indicating the upper boundary of the support of X. If not provided, it is set to the maximum value from `x`.}
}
\value{
A vector of 2 numeric entries, with the first being smaller than the second.
}
\description{
`checkBoundaries()` tests if `x` is a numeric vector with no missing values, throwing an error if it is not, and returns the boundaries of the interval if all entries of `x` lie within this interval.
}
\examples{
x <- runif(100)
checkBoundaries(x)
try(checkBoundaries(x, m = 2, M = 3))
try(checkBoundaries(x, m = 1, M = 0))
}
