% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Xunif.R
\name{f_Xunif}
\alias{f_Xunif}
\title{Standardized effects for a Uniform covariate}
\usage{
f_Xunif(x, model, K, adj_mat, m = NULL, M = NULL, fixed = TRUE, plot_check = F)
}
\arguments{
\item{x}{A numeric vector (for `model="linear","pspline1","pspline2","pspline_2D"`) or an ordered factor with more than 1 level (for `model="iid","rw1","rw2","besag"``).}

\item{model}{String to select the type of effect.}

\item{K}{A positive integer larger than 3, specifying the number of basis functions. Needed for `model="pspline1","pspline2","pspline_2D"`.}

\item{adj_mat}{Matrix needed for `model="besag"`, representing the adjacency matrix from the lattice.}

\item{m}{A number indicating the lower boundary of the support of X. If not provided, it is set to the minimum value from `x`. Needed for `model="linear","pspline1","pspline2","pspline_2D"`.}

\item{M}{A number indicating the upper boundary of the support of X. If not provided, it is set to the maximum value from `x`.  Needed for `model="linear","pspline1","pspline2","pspline_2D"`.}

\item{fixed}{Logical, indicating whether the effect must be treated as fixed (TRUE) or random (FALSE). If it is treated as fixed, a 0 mean constraint is imposed By default, TRUE. The only model affected is `model="iid"`.}

\item{plot_check}{Logical, indicating whether a plot to check the standardization must be printed (TRUE) or not (FALSE). By default, FALSE.}
}
\description{
`f_Xunif()` returns a list to build various standardized effects (specified through the `model` argument) under the assumption that the corresponding covaiate follows a Uniform distribution (either, discrete or continuous). The first three elements of the list are designed to be used in INLA to specify an effect in its standardized version through the model `generic0`. The list also reports the scaling constant, a sample from the Uniform distribution on the covariate, and  the basis matrix evaluated at this sample: these elements can be useful for the user. The argument `plot_check=TRUE` returns a plot showing realizations of the standardized effect along with the histograms of the mean and variance of each realization.
}
\examples{
#Example 1
K <- 20
x <- factor(sample(1:K, 100, replace = TRUE), ordered = TRUE, levels = c(1:K))
f_Xunif(x, model = "iid", fixed = FALSE, plot_check=TRUE)
f_Xunif(x, model = "rw1", plot_check=TRUE)
f_Xunif(x, model = "rw2", plot_check=TRUE)

#Example 2
K <- 20
x<- runif(1000,min=2,max=5)
f_Xunif(x, model = "linear", plot_check=TRUE)
f_Xunif(x, model = "pspline1",   K = K, plot_check=TRUE)
f_Xunif(x, model = "pspline2",   K = K, plot_check=TRUE)
}
