% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid.R
\name{iid_standard}
\alias{iid_standard}
\title{Standardized cluster effect}
\usage{
iid_standard(x, fixed = TRUE)
}
\arguments{
\item{x}{An ordered factor, with more than 1 level.}

\item{fixed}{Logical, indicating whether the effect must be treated as fixed (TRUE) or random (FALSE). If it is treated as fixed, a 0 mean constraint is imposed By default, TRUE.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item{`precision`: precision matrix.}
\item{`basis`:  basis matrix evaluated at `x`}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
 \item{`null_space`: a matrix, representing the null space of the precision matrix.}
 \item{`X_distribution`: a numeric vector sampled from the Uniform distribution on X.}
 \item{`basis_distribution`: basis matrix evaluated at `X_distribution`.}
}
}
\description{
Standardized cluster effect
}
\examples{
K <- 20
x <- factor(sample(1:K, 100, replace = TRUE), ordered = TRUE, levels = c(1:K))
iid_standard(x)
iid_standard(x, fixed = FALSE)
}
