% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOrderedFactor.R
\name{isOrderedFactor}
\alias{isOrderedFactor}
\title{Check if the argument is an ordered factor.}
\usage{
isOrderedFactor(x)
}
\arguments{
\item{x}{An ordered factor, with more than 1 level.}
}
\description{
\code{isOrderedFactor()} throws an error message if the argument is not an ordered factor with more than 2 levels or contains NA values.
}
\details{
This function is used in other functions of the \code{scaleGMRF} package to check whether their first argument is an ordered factor.
}
\examples{
K <- 20
x <- factor(1:K, ordered = TRUE, levels = c(1:K))

# Valid example -------------------------------------------------
isOrderedFactor(x) # NULL

# Examples throwing errors ---------------------------------------
try(isOrderedFactor(c(x, NA)))
try(isOrderedFactor(factor(1:K, levels = c(1:K))))
try(isOrderedFactor(factor(rep(1, K), levels = 1)))
}
