% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSPSD.R
\name{isSPSD}
\alias{isSPSD}
\title{Test if a Matrix is Symmetric Positive (Semi-)Definite}
\usage{
isSPSD(M)
}
\arguments{
\item{M}{A matrix object.}
}
\description{
`isSPSD()` throws an error message if the argument is not a non-empty symmetric positive (semi-)definite matrix without NA entries and not nilpotent. The error message explains which property is not respected.
}
\examples{
isSPSD(diag(10)) # NULL
try(isSPSD(diag(0))) # error
try(isSPSD(diag(NA, ncol = 3, nrow = 3))) # error
try(isSPSD(matrix(1, ncol = 3, nrow = 2))) # error
}
