% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidBasis.R
\name{isValidBasis}
\alias{isValidBasis}
\title{Test if a matrix is an appropriate basis}
\usage{
isValidBasis(D, Q)
}
\arguments{
\item{D}{a matrix.}

\item{Q}{a square matrix.}
}
\description{
`isValidBasis()` throws an error message if the first argument is not a non-empty matrix of the correct dimension. The error message explains which property is not respected.
}
\examples{
x <- seq(0, 1, length.out = 100)
isValidBasis(D = bspline(x, K = 10), Q = diag(10)) # NULL

try(isValidBasis(D = bspline(x, K = 20), Q = diag(10))) # error
}
