% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear_standard}
\alias{linear_standard}
\title{Standardized linear effect}
\usage{
linear_standard(x, m = NULL, M = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{m}{A number indicating the lower boundary of the support of X. If not provided, it is set to the minimum value from `x`.}

\item{M}{A number indicating the upper boundary of the support of X. If not provided, it is set to the maximum value from `x`.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item{`precision`: precision matrix.}
\item{`basis`:  basis matrix evaluated at `x`}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
 \item{`null_space`: a matrix, representing the null space of the precision matrix.}
 \item{`X_distribution`: a numeric vector sampled from the Uniform distribution on X.}
 \item{`basis_distribution`: basis matrix evaluated at `X_distribution`.}
}
}
\description{
`linear_standard()` provides a list of elements to build a standardized linear effect.
}
\examples{
x <- seq(0, 1, length.out = 100)

linear_standard(x)

}
