% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean0.R
\name{mean0_GMRF}
\alias{mean0_GMRF}
\title{Precision matrix of a GMRF subject to 0-mean constraint.}
\usage{
mean0_GMRF(Q, D = NULL, rank_def = NULL)
}
\arguments{
\item{Q}{A square matrix, representing the precision matrix on the coefficients.}

\item{D}{A matrix, representing the basis matrix of the effect. If not provided, it is set to an identity matrix of dimension equal to `Q`.}

\item{rank_def}{The rank deficiency of the matrix `Q`. If not provided, it is estimated.}
}
\value{
A list of 3 elements containing:
\itemize{
\item{`Q`: a matrix, representing the precision matrix under the 0-mean constraint.}
\item{`D`: a matrix, representing the basis under the 0-mean constraint.}
#' \item{`A`: a matrix, representing the 0-mean constraint as the coefficients of a linear combination.}
}
}
\description{
`mean0_GMRF` returns the precision matrix subject to a 0-mean constraint, based on the provided precision and basis matrices.
}
\examples{
mean0_GMRF(diag(10))
mean0_GMRF(as.matrix(spam::precmat.RW1(10)), rank_def = 1)
}
