% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modif_Pspline.R
\name{modif_Pspline}
\alias{modif_Pspline}
\title{Provide a modified version of the precision matrix for P-Spline effects}
\usage{
modif_Pspline(x, K, order = 2, sparse_sol = T)
}
\arguments{
\item{x}{A numeric vector, containing values between 0 and 1.}

\item{K}{A positive integer larger than 3, specifying the number of basis functions on the first dimension.}

\item{order}{Either 1 or 2, specifying the order of the random walk process on the coefficients.}

\item{sparse_sol}{logical indicating whether the solution is to be provided in the sparse version. Otherwise, the solution is provided in a non-sparse format. By default, TRUE.}
}
\value{
A list of 4 elements, containing:
\itemize{
\item{`basis`: a matrix, representing the basis matrix.}
\item{`prec`: a matrix, representing the precision matrix.}
\item{`null_space`: a matrix, representing the null space of the precision matrix.}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
}
}
\description{
`modif_Pspline()` provides a list of elements to build the modified version of a P-Spline effect, either of order 1 or 2. Specifically, the appropriate basis matrix, precision matrix, null space, and scaling constant are provided.
}
\examples{
x <- seq(0, 1, length.out = 100)

modif_Pspline(x, K = 20,order = 2)

}
