% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modif_Pspline2D.R
\name{modif_Pspline2D}
\alias{modif_Pspline2D}
\title{Provide a modified version of the precision matrix for P-Spline effects}
\usage{
modif_Pspline2D(x, K1, K2, sparse_sol = T)
}
\arguments{
\item{x}{A matrix of 2 columns, containing values between 0 and 1.}

\item{K1}{A positive integer larger than 3, specifying the number of basis functions on the first dimension.}

\item{K2}{A positive integer larger than 3, specifying the number of basis functions on the second dimension.}

\item{sparse_sol}{logical indicating whether the solution is to be provided in the sparse version. Otherwise, the solution is provided in a non-sparse format. By default, TRUE.}
}
\value{
A list of 4 elements, containing:
\itemize{
\item{`basis`: a matrix, representing the basis matrix.}
\item{`prec`: a matrix, representing the precision matrix.}
\item{`null_space`: a matrix, representing the null space of the precision matrix.}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
}
}
\description{
`modif_Pspline2D()` provides a list of elements to build the modified version of a 2-dimensional P-Spline effect. Specifically, the appropriate basis matrix, precision matrix, null space, and scaling constant are provided.
}
\examples{
x_grid <- as.matrix(expand.grid(
  seq(0, 1, length.out = 100),
  seq(0, 1, length.out = 100)
))
modif_Pspline2D(x_grid, K1 = 5, K2 = 5)
}
