% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean0_constr.R
\name{prec_under_0mean}
\alias{prec_under_0mean}
\title{Precision matrix of a GMRF subject to 0-mean constraint.}
\usage{
prec_under_0mean(Q, D = NULL, rank_def = NULL)
}
\arguments{
\item{Q}{A square matrix, representing the precision matrix on the coefficients.}

\item{D}{A matrix, representing the basis matrix of the effect. If not provided, it is set to an identity matrix of dimension equal to `Q`.}

\item{rank_def}{The rank deficiency of the matrix `Q`. If not provided, it is estimated.}
}
\value{
A list of 2 elements containing:
\itemize{
\item{`Q_constr`: a matrix, representing the precision matrix under the 0-mean constraint.}
\item{`constr`: a numeric vector, representing the 0-mean constraint.}
}
}
\description{
`prec_under_0mean` returns the precision matrix subject to a 0-mean constraint, based on the provided precision and basis matrices.
}
\examples{
prec_under_0mean(diag(10))
prec_under_0mean(as.matrix(spam::precmat.RW1(10)), rank_def = 1)
}
