% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pspline1.R, R/Pspline2.R
\name{Pspline1_standard}
\alias{Pspline1_standard}
\alias{Pspline2_standard}
\title{Standardized P-Spline effects}
\usage{
Pspline1_standard(x, K, m = NULL, M = NULL, sparse_sol = T)

Pspline2_standard(x, K, m = NULL, M = NULL, sparse_sol = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{K}{A positive integer larger than 3, specifying the number of basis functions.}

\item{m}{A number indicating the lower boundary of the support of X. If not provided, it is set to the minimum value from `x`.}

\item{M}{A number indicating the upper boundary of the support of X. If not provided, it is set to the maximum value from `x`.}

\item{sparse_sol}{Logical, indicating whether the solution is to be provided in the sparse version. Otherwise, the solution is provided in a non-sparse format. By default, TRUE.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item{`precision`: precision matrix.}
\item{`basis`:  basis matrix evaluated at `x`}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
 \item{`null_space`: a matrix, representing the null space of the precision matrix.}
 \item{`X_distribution`: a numeric vector sampled from the Uniform distribution on X.}
 \item{`basis_distribution`: basis matrix evaluated at `X_distribution`.}
}
}
\description{
`Pspline1_standard()` provides a list of elements to build a standardized P-Spline effect with a random walk process of order 1 on the coefficients. Note that the modified version is used, as detailed in ???.

`Pspline2_standard()` provides a list of elements to build a standardized P-Spline effect with a random walk process of order 2 on the coefficients. Note that the modified version is used, as detailed in ???.
}
\examples{
x <- seq(0, 1, length.out = 100)

Pspline1_standard(x, K = 10)
Pspline1_standard(x, K = 20)
Pspline2_standard(x, K = 10)
Pspline2_standard(x, K = 20)

}
