% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspline.R
\name{pspline_standard}
\alias{pspline_standard}
\title{Standardized P-Spline effects}
\usage{
pspline_standard(x, K, order, m = NULL, M = NULL, sparse_sol = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{K}{A positive integer larger than 3, specifying the number of basis functions.}

\item{order}{Either 1 or 2, representing the order of the random walk.}

\item{m, M}{A pair of numbers indicating respectively the lower and upper boundaries of the support of X. If not provided, it is set to the extreme values of \code{x}.}

\item{sparse_sol}{Logical, indicating whether the solution is to be provided in the sparse version. Otherwise, the solution is provided in a non-sparse format. By default, TRUE.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item \code{precision}: precision matrix
\item \code{basis}:  basis matrix evaluated at \code{x}
\item \code{scaling_constant}: a positive number, representing the appropriate scaling constant
\item \code{null_space}: a matrix, representing the null space of the precision matrix
\item \code{X_distribution}: a numeric vector of values sampled from the Uniform distribution of X
\item \code{basis_distribution}: basis matrix evaluated at \code{X_distribution}
}
}
\description{
\code{pspline_standard()} provides a list of elements to build a standardized P-Spline effect with a random walk process of order 1 or 2 on the coefficients. Note that the modified version is used, as detailed in ???.
}
\examples{
x <- seq(0, 1, length.out = 100)

pspline_standard(x, K = 10, order = 1)
pspline_standard(x, K = 20, order = 1)
pspline_standard(x, K = 10, order = 2)
pspline_standard(x, K = 20, order = 2)

}
