% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_GMRF.R
\name{r_GMRF}
\alias{r_GMRF}
\title{Generate realizations of a GMRF}
\usage{
r_GMRF(Q, D = NULL, sigma2 = 1, rank_def = NULL, n_sim = 100)
}
\arguments{
\item{Q}{A matrix, representing the precision matrix on the coefficients of the Gaussian effect.}

\item{D}{A matrix, representing the basis of the effect. If not provided, it is set to an identity matrix of dimension equal to \code{Q}.}

\item{sigma2}{Positive number, indicating the value of the variance parameter. By default, 1.}

\item{rank_def}{A non-negative integer, indicating the rank deficiency of the matrix \code{Q}. If not provided, it is estimated.}

\item{n_sim}{Positive integer, indicating the number of realizations to generate. By default, \code{n_sim=100}.}
}
\value{
A numeric matrix, containing a realization of the process in each column.
}
\description{
\code{r_GMRF()} returns a matrix where each column is a realization of the specified Gaussian effect.
}
\details{
This function is used in the \code{standardize_GMRF()} and \code{f_Xunif()} functions of the \code{scaleGMRF} package to generate realizations of Gaussian effects.
}
\examples{
# Example for an i.i.d. effect --------------------
r_GMRF(Q = diag(10))

# Example for a linear effect ------------------------
r_GMRF(Q = matrix(1), D = matrix(runif(100)))

# Example for a random walk effect of order 1 --------
r_GMRF(Q = as.matrix(spam::precmat.RW1(10)), rank_def = 1)

}
