% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw_standard}
\alias{rw_standard}
\title{Standardized random walk effects}
\usage{
rw_standard(x, order)
}
\arguments{
\item{x}{An ordered factor, with more than 1 level.}

\item{order}{Either 1 or 2, representing the order of the random walk.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item{`precision`: precision matrix.}
\item{`basis`:  basis matrix evaluated at `x`}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
 \item{`null_space`: a matrix, representing the null space of the precision matrix.}
 \item{`X_distribution`: a numeric vector sampled from the Uniform distribution on X.}
 \item{`basis_distribution`: basis matrix evaluated at `X_distribution`.}
}
}
\description{
`rw_standard()` provides a list of elements to build a standardized random walk effect of order 1 or 2.
}
\examples{
K <- 20
x <- factor(sample(1:K, 100, replace = TRUE), ordered = TRUE, levels = c(1:K))

rw_standard(x, order = 1)
rw_standard(x, order = 2)
}
