% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_GMRF}
\alias{standardize_GMRF}
\title{Standardize a GMRF}
\usage{
standardize_GMRF(
  Q,
  D = NULL,
  rank_def = NULL,
  fixed = TRUE,
  plot_check = FALSE,
  n_sim = 1000
)
}
\arguments{
\item{Q}{A square matrix, representing the precision matrix on the coefficients.}

\item{D}{A matrix, representing the basis matrix of the effect. If not provided, it is set to an identity matrix of dimension equal to `Q`.}

\item{rank_def}{The rank deficiency of the matrix `Q`. If not provided, it is estimated.}

\item{fixed}{logical indicating whether the effect is to be considered fixed.}

\item{plot_check}{Logical, indicating whether a plot to check the standardization must be printed (TRUE) or not (FALSE). By default, FALSE.}

\item{n_sim}{Positive integer, indicating the number of realizations to generate for the check plot.}
}
\value{
A list of 3 or 4 elements containing:
\itemize{
\item{`Q`: a matrix, representing the precision matrix under the 0-mean constraint.}
\item{`D`: a matrix, representing the basis under the 0-mean constraint.}
\item{`C`: a number, representing the scaling constant.}
\item{`A`: a matrix, representing the 0-mean constraint as the coefficients of a linear combination.}
}
}
\description{
`scale_GMRF()` returns the appropriate scaling constant for a given effect of an LGM model, defined through a basis matrix and a precision matrix.
}
\examples{
prec <- as.matrix(spam::precmat.RW1(10))
standardize_GMRF(prec, rank_def = 1)
standardize_GMRF(prec, rank_def = 1, fixed = FALSE)
standardize_GMRF(diag(10))
standardize_GMRF(diag(10), fixed = FALSE)
}
