% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBoundaries.R
\name{xBoundaries}
\alias{xBoundaries}
\title{Valid boundaries for a numeric vector}
\usage{
xBoundaries(x, m = NULL, M = NULL)
}
\arguments{
\item{x}{A numeric vector, indicating the values on which the basis must be evaluated.}

\item{m, M}{A pair of numbers indicating respectively the lower and upper boundary of the support of X. If not provided, they are set to the minimum and maximum values of \code{x}.}
}
\value{
A numeric vector with 2 elements, the first being smaller than the second.
}
\description{
\code{xBoundaries()} returns valid boundaries for the specified numeric vector.
}
\details{
This function is used in other functions of the \code{scaleGMRF} package to return appropriate boundaries for a numeric vector. If the boundaries are specified, the function simply checks that each entry of the numeric vector lies between the two extremes. Otherwise, the function first estimates the boundaries using the minimum and maximum values of the vector.
}
\examples{
# Valid example --------------------------------------------------
x <- runif(100)
xBoundaries(x)

# Examples throwing errors ---------------------------------------
try(xBoundaries(x, m = 2, M = 3))
try(xBoundaries(x, m = 1, M = 0))
}
