% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline}
\alias{bspline}
\title{Create a basis of cubic uniform B-Spline functions}
\usage{
bspline(x, K)
}
\arguments{
\item{x}{A numeric vector, with values contained between 0 and 1.}

\item{K}{A positive integer larger than 3, specifying the number of basis functions.}
}
\value{
A matrix with `K` columns and a number of rows equal to the length of `x`.
}
\description{
`bspline()` creates a basis matrix with `K` columns, each of them evaluating the numeric vector `x` at a cubic B-Spline function defined on equidistant nodes on the [0,1] interval.
}
\examples{
x <- seq(0, 1, length.out = 100)
bspline(x, K = 10)
}
