% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline2D.R
\name{bspline_2D}
\alias{bspline_2D}
\title{Create a basis of equidistant cubic 2D B-Spline functions}
\usage{
bspline_2D(x, K, m = NULL, M = NULL)
}
\arguments{
\item{x}{A numeric matrix with 2 columns, indicating the values on which the basis must be evaluated.}

\item{K}{Two positive integers larger than 3, indicating the number of basis functions for each dimension.}

\item{m, M}{Two pairs of numbers, indicating respectively the lower and upper boundaries of the rectangular support on which to define the basis functions. If not provided, they are estimated using the minimum and maximum values from the columns of \code{x}.}
}
\value{
A matrix with \code{K[1]}x\code{K[2]} columns and a number of rows equal to the one of \code{x}.
}
\description{
\code{bspline_2D()} creates a basis matrix with \code{K[1]}x\code{K[2]} columns, each of them evaluating the values in \code{x} at a bivariate cubic B-Spline function, defined as the Kronecker product between 2 univariate B-Spline functions.
}
\details{
This function is used in the \code{pspline_2D_standard()} function of the \code{scaleGMRF} package to build a basis of equidistant cubic 2D B-Splines on a rectangular support.
}
\examples{
x_grid <- as.matrix(expand.grid(
  seq(0, 1, length.out = 100),
  seq(0, 1, length.out = 100)
))

bspline_2D(x_grid, K = c(5, 5))

}
