% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_GMRF}
\alias{check_GMRF}
\title{Plots to check standardization of a GMRF.}
\usage{
check_GMRF(realizations, fixed = T, X_dist = NULL)
}
\arguments{
\item{realizations}{A numeric matrix, containing a realization of the process in each column.}

\item{fixed}{Logical, indicating whether the effect must be treated as fixed or random. By default, \code{TRUE}.}

\item{X_dist}{A numeric matrix, containing values sampled from the distribution of X. By default, \code{NULL}.}
}
\value{
Plot.
}
\description{
\code{check_GMRF()} returns a plot made up by 1 or more panels to check the standardization of a Gaussian effect. If a sample of the distribution of the covariate is specified, a plot of the realizations is also returned.
}
\details{
This function is used to visually check the standardization of the Gaussian effect. It is called in the functions \code{standardize_GMRF()} and \code{standardize_X_unif()} of the \code{scaleGMRF} package when \code{plot_check=T}. The plot provided in the output is different for fixed and random effects: in both cases, the "Mean" value in the plot should be close to 1 if standardization has been correctly applied. See \code{vignette("standardization",package="scaleGMRF")} for more details.
}
