% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_GMRF.R
\name{check_GMRF}
\alias{check_GMRF}
\title{Checks if the effect has been properly standardized.}
\usage{
check_GMRF(result)
}
\arguments{
\item{result}{A list returned from the `f_Xunif()` function.}
}
\value{
Plot.
}
\description{
`check_GMRF()` returns a plot with three panels. Realizations of the effect from the argument are generated and displayed in the first panel. The expected value and variances of the realizations with respect to the Uniform distribution of the covariate are computed and plotted via histograms. The means of both statistics are also reported, which should be equal respectively to 0 (only if `fixed=TRUE`) and 1.
}
