% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_inv.R
\name{gen_inv}
\alias{gen_inv}
\title{Generalized Inverse of a Matrix}
\usage{
gen_inv(Q, rank_def = NULL)
}
\arguments{
\item{Q}{An SPSD matrix.}

\item{rank_def}{A non-negative integer, indicating the rank deficiency of the matrix. If \code{NULL}, the rank deficiency is estimated.}
}
\description{
\code{gen_inv()} calculates the Moore-Penrose generalized inverse of a Symmetric Positive (Semi-)Definite (SPSD) matrix, with the possibility of explicitly stating the rank-deficiency when the matrix is positive semi-definite.
}
\details{
This function is used in other functions of the \code{scaleGMRF} package to compute the inverse of rank-deficient precision matrices, for which the rank-deficiency is known and does not need to be estimated, as it happens in the function \code{\link[MASS:ginv]{MASS::ginv()}}.
}
\examples{
# Example with a diagonal matrix -----------------------------------
gen_inv(diag(10), rank_def = 0)

# Example with a random walk of order 1 precision matrix ------------
gen_inv(as.matrix(spam::precmat.RW1(10)), rank_def = 1)

# Example with a random walk of order 2 precision matrix ------------
gen_inv(as.matrix(spam::precmat.RW2(10)), rank_def = 2)

}
\seealso{
\code{\link[=solve]{solve()}},\code{\link[MASS:ginv]{MASS::ginv()}}
}
