% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid.R
\name{iid_standard}
\alias{iid_standard}
\title{Standardized cluster effect}
\usage{
iid_standard(x, fixed = TRUE)
}
\arguments{
\item{x}{An ordered factor, with more than 1 level.}

\item{fixed}{Logical, indicating whether the effect must be treated as fixed (TRUE) or random (FALSE). If it is treated as fixed, a 0 mean constraint is imposed By default, TRUE.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item \code{Q}: precision matrix
\item \code{D}: basis matrix evaluated at \code{x}
\item \code{C}: a positive number, representing the appropriate scaling constant
\item \code{null_space}: a matrix, representing the null space of the precision matrix
\item \code{X_distribution}: a numeric vector (or matrix for \code{pspline2D}) of values sampled from the Uniform distribution of X
\item \code{basis_distribution}: basis matrix evaluated at \code{X_distribution}
}
}
\description{
\code{iid_standard()} provides a list of elements to build a standardized cluster effect.
}
\examples{
K <- 20
x <- factor(sample(1:K, 100, replace = TRUE), ordered = TRUE, levels = c(1:K))
#Example as fixed effect ------------
iid_standard(x)

#Example as random effect ------------
iid_standard(x, fixed = FALSE)
}
