% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOrderedFactor.R
\name{isOrderedFactor}
\alias{isOrderedFactor}
\title{Check if the argument is an ordered factor.}
\usage{
isOrderedFactor(x)
}
\arguments{
\item{x}{An ordered factor, with more than 1 level.}
}
\description{
`isOrderedFactor()` throws an error message if the argument is not an ordered factor with more than 2 levels or contains NA values.
}
\examples{
K <- 20
x <- factor(1:K, ordered = TRUE, levels = c(1:K))
isOrderedFactor(x) # NULL
try(isOrderedFactor(c(x,NA))) # error
try(isOrderedFactor(factor(1:K, levels = c(1:K)))) # error
try(isOrderedFactor(factor(rep(1,K), levels = 1))) # error
}
