% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSPSD.R
\name{isSPSD}
\alias{isSPSD}
\title{Check if a Matrix is Symmetric Positive (Semi-)Definite}
\usage{
isSPSD(Q)
}
\arguments{
\item{Q}{A matrix object.}
}
\description{
\code{isSPSD()} throws an error message if the argument is not a non-empty Symmetric Positive (Semi-)Definite (SPSD) matrix, without missing values, and not nilpotent. The error message explains which property is not respected.
}
\details{
This function is used in other functions of the \code{scaleGMRF} package to check whether a matrix is a valid precision matrix.
}
\examples{
# Example with SPSD matrix ---------------------------------
isSPSD(diag(10)) # NULL

# Example with empty matrix ---------------------------------
try(isSPSD(diag(0))) # error

# Example with missing values ---------------------------------
try(isSPSD(diag(NA, ncol = 3, nrow = 3))) # error

# Example with non-square matrix ---------------------------------
try(isSPSD(matrix(1, ncol = 3, nrow = 2))) # error

}
