% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear_standard}
\alias{linear_standard}
\title{Standardized linear effect}
\usage{
linear_standard(x, m = NULL, M = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{m, M}{A pair of numbers indicating respectively the lower and upper boundaries of the support of X. If not provided, it is set to the extreme values of \code{x}.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item \code{precision}: precision matrix
\item \code{basis}:  basis matrix evaluated at \code{x}
\item \code{scaling_constant}: a positive number, representing the appropriate scaling constant
\item \code{null_space}: a matrix, representing the null space of the precision matrix
\item \code{X_distribution}: a numeric vector of values sampled from the Uniform distribution of X
\item \code{basis_distribution}: basis matrix evaluated at \code{X_distribution}
}
}
\description{
\code{linear_standard()} provides a list of elements to build a standardized linear effect.
}
\examples{
x <- seq(0, 1, length.out = 100)

linear_standard(x)

}
