% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean0.R
\name{mean0_GMRF}
\alias{mean0_GMRF}
\title{Precision matrix of a GMRF subject to 0-mean constraint.}
\usage{
mean0_GMRF(Q, D = NULL, rank_def = NULL)
}
\arguments{
\item{Q}{A square matrix, representing the precision matrix on the coefficients of the Gaussian effect.}

\item{D}{A matrix, representing the basis matrix of the effect. If not provided, it is set to an identity matrix of dimension equal to \code{Q}.}

\item{rank_def}{The rank deficiency of the matrix \code{Q}. If not provided, it is estimated.}
}
\value{
A list of 3 elements containing:
\itemize{
\item \code{Q}: a matrix, representing the precision matrix subject to the 0-mean constraint.
\item \code{D}: a matrix, representing the basis subject to the 0-mean
\item \code{A}: a matrix with 1 column, representing the 0-mean constraint as the coefficients of a linear combination.
}
}
\description{
\code{mean0_GMRF} returns the precision matrix subject to a 0-mean constraint of a Gaussian effect, defined through a basis matrix and a set of Normally distributed coefficients with null mean and given precision matrix.
}
\examples{
mean0_GMRF(diag(10))
mean0_GMRF(as.matrix(spam::precmat.RW1(10)), rank_def = 1)
}
