% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspline_2D.R
\name{pspline_2D_standard}
\alias{pspline_2D_standard}
\title{Standardized 2D P-Spline effect}
\usage{
pspline_2D_standard(x, K, m = NULL, M = NULL, sparse_sol = T)
}
\arguments{
\item{x}{A matrix of 2 columns of numeric entries.}

\item{K}{A vector containing 2 positive integers larger than 3, specifying the number of basis functions respectively for the first and second dimensions.}

\item{m}{A vector of 2 numbers, indicating the lower extremes in the two dimensions. If not provided, it is set to the minimum values from the `x` columns.}

\item{M}{A vector of 2 numbers, indicating the upper extremes in the two dimensions. If not provided, it is set to the maximum values from the `x` columns.}

\item{sparse_sol}{Logical, indicating whether the solution is to be provided in its sparse version. Otherwise, the solution is provided in a non-sparse format. By default, TRUE.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item{`precision`: precision matrix.}
\item{`basis`:  basis matrix evaluated at `x`}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
 \item{`null_space`: a matrix, representing the null space of the precision matrix.}
 \item{`x_dist`: a matrix of 2 columns representing the Uniform distribution assumed on X1, X2.}
 \item{`basis_dist`: basis matrix evaluated at `x_dist`.}
}
}
\description{
`pspline_2D_standard()` provides a list of elements to build a standardized 2-dimensional P-Spline effect. Specifically, the scaled precision matrix, the basis matrix, the null space, and the scaling constant are provided.
}
\examples{
x_grid <- as.matrix(expand.grid(
  seq(0, 1, length.out = 100),
  seq(0, 1, length.out = 100)
))
pspline_2D_standard(x_grid, K = c(5, 5))
}
