% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspline.R
\name{pspline_standard}
\alias{pspline_standard}
\title{Standardized P-Spline effects}
\usage{
pspline_standard(x, K, order, m = NULL, M = NULL, sparse_sol = T)
}
\arguments{
\item{x}{A numeric vector.
@param K A positive integer larger than 3, specifying the number of basis functions.
@param order Either 1 or 2, representing the order of the random walk.}

\item{m}{A number indicating the lower boundary of the support of X. If not provided, it is set to the minimum value from `x`.}

\item{M}{A number indicating the upper boundary of the support of X. If not provided, it is set to the maximum value from `x`.}

\item{sparse_sol}{Logical, indicating whether the solution is to be provided in the sparse version. Otherwise, the solution is provided in a non-sparse format. By default, TRUE.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item{`precision`: precision matrix.}
\item{`basis`:  basis matrix evaluated at `x`}
\item{`scaling_constant`: a positive number, representing the appropriate scaling constant.}
 \item{`null_space`: a matrix, representing the null space of the precision matrix.}
 \item{`X_distribution`: a numeric vector sampled from the Uniform distribution on X.}
 \item{`basis_distribution`: basis matrix evaluated at `X_distribution`.}
}
}
\description{
`pspline_standard()` provides a list of elements to build a standardized P-Spline effect with a random walk process of order 1 or 2 on the coefficients. Note that the modified version is used, as detailed in ???.
}
\examples{
x <- seq(0, 1, length.out = 100)

pspline_standard(x, K = 10, order = 1)
pspline_standard(x, K = 20, order = 1)
pspline_standard(x, K = 10, order = 2)
pspline_standard(x, K = 20, order = 2)

}
