% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw_standard}
\alias{rw_standard}
\title{Standardized random walk effects}
\usage{
rw_standard(x, order)
}
\arguments{
\item{x}{An ordered factor, with more than 1 level.}

\item{order}{Either 1 or 2, representing the order of the random walk.}
}
\value{
A list of 6 elements, containing:
\itemize{
\item \code{precision}: precision matrix
\item \code{basis}:  basis matrix evaluated at \code{x}
\item \code{scaling_constant}: a positive number, representing the appropriate scaling constant
\item \code{null_space}: a matrix, representing the null space of the precision matrix
\item \code{X_distribution}: a numeric vector (or matrix) of values sampled from the Uniform distribution of X
\item \code{basis_distribution}: basis matrix evaluated at \code{X_distribution}
}
}
\description{
\code{rw_standard()} provides a list of elements to build a standardized random walk effect of order 1 or 2.
}
\examples{
K <- 20
x <- factor(sample(1:K, 100, replace = TRUE), ordered = TRUE, levels = c(1:K))

# Example for a random walk of order 1 --------------------
rw_standard(x, order = 1)

# Example for a random walk of order 2 --------------------
rw_standard(x, order = 2)
}
