% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_GMRF}
\alias{scale_GMRF}
\title{Expectation-based scaling for an LGM effect}
\usage{
scale_GMRF(Q, D = NULL, rank_def = NULL)
}
\arguments{
\item{Q}{A square matrix, representing the precision matrix on the coefficients.}

\item{D}{A matrix, representing the basis matrix of the effect. If not provided, it is set to an identity matrix of dimension equal to `Q`.}

\item{rank_def}{The rank deficiency of the matrix `Q`. If not provided, it is estimated.}
}
\value{
A positive number to be used for scaling.
}
\description{
`scale_GMRF()` returns the appropriate scaling constant for a given effect of an LGM model, defined through a basis matrix and a precision matrix.
}
\examples{
scale_GMRF(Q = diag(10))
prec <- as.matrix(spam::precmat.RW1(10))
scale_GMRF(Q = prec, rank_def = 1)
}
