% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_GMRF}
\alias{standardize_GMRF}
\title{Standardization of a GMRF}
\usage{
standardize_GMRF(
  Q,
  D = NULL,
  rank_def = NULL,
  fixed = TRUE,
  scale_Q = TRUE,
  plot_check = FALSE,
  n_sim = 100
)
}
\arguments{
\item{Q}{A matrix, representing the precision matrix on the coefficients of the Gaussian effect.}

\item{D}{A matrix, representing the basis of the effect. If not provided, it is set to an identity matrix of dimension equal to \code{Q}.}

\item{rank_def}{A non-negative integer, indicating the rank deficiency of the matrix \code{Q}. If not provided, it is estimated.}

\item{fixed}{Logical, indicating whether the effect must be treated as fixed or random. An effect is always considered fixed if it only has one basis function. By default, \code{TRUE}.}

\item{scale_Q}{Logical, indicating whether the precision matrix must be scaled or not. By default, \code{TRUE}.}

\item{plot_check}{Logical, indicating whether a plot to check the standardization must be printed or not. By default, \code{FALSE}.}

\item{n_sim}{Positive integer, indicating the number of realizations to generate for the check plot. By default, \code{n_sim=100}.}
}
\value{
A list of 4 elements containing:
\itemize{
\item \code{Q}: a matrix, representing the precision matrix.
\item \code{D}: a matrix, representing the basis.
\item \code{C}: a positive number, representing the scaling constant.
\item \code{A}: a matrix with 1 column, representing the 0-mean constraint as the coefficients of a linear combination. \code{NULL} for \code{fixed=FALSE}.
}
}
\description{
\code{standardize_GMRF()} standardizes a Gaussian effect and returns the elements needed to specify the standardized effect in INLA.
}
\details{
This function is one of the two main functions of the \code{scaleGMRF} package. More details about this function can be found in \code{vignette("standardization",package="scaleGMRF")}.
}
\examples{

# Example for an i.i.d effect ------------------------
prec <- diag(10)
# fixed effect
standardize_GMRF(prec)
# random effect
standardize_GMRF(prec, fixed = FALSE)

# Example for a linear effect ------------------------
basis <- matrix(runif(100))
# fixed effect
standardize_GMRF(Q = matrix(1), D = basis)

# Example for a random walk effect of order 1 ------------------------
prec <- as.matrix(spam::precmat.RW1(10))
# fixed effect
standardize_GMRF(prec, rank_def = 1)
# random effect
standardize_GMRF(prec, rank_def = 1, fixed = FALSE)

}
