% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_GMRF}
\alias{standardize_GMRF}
\title{Standardize a GMRF}
\usage{
standardize_GMRF(Q, D = NULL, rank_def = NULL, fixed = FALSE)
}
\arguments{
\item{Q}{A square matrix, representing the precision matrix on the coefficients.}

\item{D}{A matrix, representing the basis matrix of the effect. If not provided, it is set to an identity matrix of dimension equal to `Q`.}

\item{rank_def}{The rank deficiency of the matrix `Q`. If not provided, it is estimated.}

\item{fixed}{logical indicating whether the effect is to be considered fixed.}
}
\value{
A square matrix representing the precision matrix after appropriate scaling.
}
\description{
`ex_scale()` returns the appropriate scaling constant for a given effect of an LGM model, defined through a basis matrix and a precision matrix.
}
\examples{
prec <- as.matrix(spam::precmat.RW1(10))
standardize_GMRF(prec, rank_def=1)
standardize_GMRF(prec, rank_def=1, fixed=T)
standardize_GMRF(diag(10))
standardize_GMRF(diag(10), fixed=T)
}
